iwm_task1/requirements.txt
watchdog==0.9.0
numba==0.46.0
tqdm==4.36.1
scikit_image==0.15.0
numpy==1.17.2
pydicom==1.4.2
Unidecode==1.1.1
matplotlib==3.1.1
scipy==1.3.1
ipython==7.13.0
Pillow==7.0.0
skimage==0.0
